Add table visualization
* Made general fetchStops line for all bus/max lines

* addsNames

Adds a button that allows you to see the names of all of the routes

* moved functions to new js

Moved table visualization functions out of the fetch.js and into tablefunctions.js

* moving tablefunctions

I moved the tablefunctions out of fetch

* THIS! works hehe

Max grab this

* merged james's code

* Got the names buttons working

Got all of the "names" buttons to work. This took forever. You must add the event listener to the object AFTER it is created. Since our table is made well after the program starts, we must do it immediately after creating of the button.

* Style changes

Updated style of visualization table

* hover buttons

added hover to some buttons

* active

added active feature to green buttons

* Comments

Added comments to sideBar CSS and HTML elements

---------

Co-authored-by: clarkm25 <clarkm25@up.edu>