James time table (#37)
* Set up a tab for a route's arrival time at a stop.

* Added Arrival Buttons

Added buttons that show the arrival times for a given stop. The arrival times are shown, alongside the bus line that is arriving at that time. When an arrival button is clicked on, it shows the route that it is attached to.

* Vincent pull req modifications

James made modifications

* minor change

here u go vincent i hope ur happy with yourself for this. look at what you've done.

---------

Co-authored-by: clarkm25 <clarkm25@up.edu>